+++ redwood-protocol-host/src/commonTest/kotlin/app/cash/redwood/protocol/host/HostProtocolAdapterTest.kt
@@ -36,6 +36,7 @@ import app.cash.redwood.widget.MutableListChildren
 import assertk.assertFailure
 import assertk.assertThat
 import assertk.assertions.hasMessage
+import assertk.assertions.hasSize
 import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
 import assertk.assertions.message
@@ -332,5 +333,49 @@ class HostProtocolAdapterTest {
         ),
       ),
     )
+  }
+
+  @Test fun oldRootChildrenTag() {
+    val hostAdapter = HostProtocolAdapter(
+      guestVersion = guestRedwoodVersion,
+      container = MutableListChildren(),
+      protocol = TestSchemaHostProtocol.create(),
+      widgetSystem = TestSchemaWidgetSystem(
+        TestSchema = TestSchemaTestingWidgetFactory(),
+        RedwoodUiBasic = RedwoodUiBasicTestingWidgetFactory(),
+        RedwoodLayout = RedwoodLayoutTestingWidgetFactory(),
+        RedwoodLazyLayout = RedwoodLazyLayoutTestingWidgetFactory(),
+      ),
+      eventSink = ::error,
+      leakDetector = LeakDetector.none(),
+    )
+
+    // Add a button.
+    hostAdapter.sendChanges(
+      listOf(
+        Create(
+          id = Id(1),
+          // Button
+          tag = WidgetTag(4),
+        ),
+        // Set Button's required color property.
+        PropertyChange(
+          id = Id(1),
+          widgetTag = WidgetTag(4),
+          propertyTag = PropertyTag(3),
+          value = JsonPrimitive(0),
+        ),
+        Add(
+          id = Id.Root,
+          // This is the old value for the root's children:
+          tag = ChildrenTag(1),
+          childId = Id(1),
+          index = 0,
+        ),
+      ),
+    )
+
+    val children = hostAdapter.node(Id.Root).children(ChildrenTag.Root)!!.nodes
+    assertThat(children).hasSize(1)
   }
 }